Sharing code between n3h modes
